requirements.txt
annotated-types==0.7.0
anyio==4.8.0
apify_client==1.9.0
apify_shared==1.2.1
av==14.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
faster-whisper==1.1.1
ffmpeg-python==0.2.0
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.2.0
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
insta-scrape==2.1.2
instascrap==1.1
Jinja2==3.1.5
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
ollama==0.4.7
onnxruntime==1.20.1
packaging==24.2
protobuf==3.20.3
pybind11==2.11.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==21.10
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
setuptools==75.8.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
yt-dlp==2025.1.26
